pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "highway-perception-allweather"
version = "0.1.0"
description = "All-weather highway perception MVP with detection, tracking, lane estimation, and condition-aware evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.2.0",
  "numpy>=1.26.0",
  "opencv-python>=4.9.0.80",
  "pyyaml>=6.0.1",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
metrics = [
  "motmetrics>=1.4.0",
  "scikit-learn>=1.4.0",
]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
